Move author details
